+++ docs/sources/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog ![](img/logos/3eiy_120.png)
 
+#### 0.1.3 (2015-11-23)
+
+- Exception handling in tests if PDB goes down (which just happened)
+- Added a separate ANISOU engine to handle those records correctly
+
+
 #### 0.1.2 (2015-11-23)
 
 - First Release